Create new Purchases class for customEntitlementComputation flavor and also extension functions
